moves test utils to subgraph module out of tests directory
<!-- FED-496 -->
