+++ apollo-federation/src/schema/validators/context.rs
@@ -99,37 +99,8 @@ impl ContextValidator for DenyInvalidContextName {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::subgraph::SubgraphError;
-    use crate::subgraph::typestate::Expanded;
-    use crate::subgraph::typestate::Subgraph;
+    use crate::subgraph::test_utils::build_and_expand;
 
-    enum BuildOption {
-        AsIs,
-        AsFed2,
-    }
-
-    fn build_inner(
-        schema_str: &str,
-        build_option: BuildOption,
-    ) -> Result<Subgraph<Expanded>, SubgraphError> {
-        let name = "S";
-        let subgraph =
-            Subgraph::parse(name, &format!("http://{name}"), schema_str).expect("valid schema");
-        let subgraph = if matches!(build_option, BuildOption::AsFed2) {
-            subgraph
-                .into_fed2_subgraph()
-                .map_err(|e| SubgraphError::new(name, e))?
-        } else {
-            subgraph
-        };
-        Ok(subgraph
-            .expand_links()
-            .map_err(|e| SubgraphError::new(name, e))?)
-    }
-
-    pub(crate) fn build_and_expand(schema_str: &str) -> Subgraph<Expanded> {
-        build_inner(schema_str, BuildOption::AsIs).expect("expanded subgraph to be valid")
-    }
     #[test]
     fn deny_underscore_in_context_name() {
         let mut errors = MultipleFederationErrors::new();
