+++ apollo-federation/tests/subgraph/mod.rs
@@ -1,3 +1,2 @@
 mod parse_expand_tests;
-pub(crate) mod subgraph_utils;
 mod subgraph_validation_tests;
